refactor: move home screen to be a composable
