+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">KMP Movies</string>
+    <string name="back">Go back</string>
+</resources>
